[fix] Remove unnecessary uint8_t usage (#290)

+++ drake_ros_examples/examples/multirobot/multirobot.cc
@@ -74,12 +74,12 @@ int main(int argc, char** argv) {
   const std::string model_name = "kuka_iiwa";
 
   // Create a 5x5 array of manipulators
-  size_t kNumRows = 5;
-  size_t kNumCols = 5;
+  int kNumRows = 5;
+  int kNumCols = 5;
   std::vector<std::vector<drake::multibody::ModelInstanceIndex>> models;
-  for (uint8_t xx = 0; xx < kNumRows; ++xx) {
+  for (int xx = 0; xx < kNumRows; ++xx) {
     std::vector<drake::multibody::ModelInstanceIndex> models_xx;
-    for (uint8_t yy = 0; yy < kNumCols; ++yy) {
+    for (int yy = 0; yy < kNumCols; ++yy) {
       // Load the model from the file and give it a name based on its X and Y
       // coordinates in the array
       std::stringstream model_instance_name;
@@ -102,8 +102,8 @@ int main(int argc, char** argv) {
   plant.Finalize();
 
   // Set the control input of each robot to uncontrolled
-  for (size_t xx = 0; xx < kNumRows; ++xx) {
-    for (size_t yy = 0; yy < kNumCols; ++yy) {
+  for (int xx = 0; xx < kNumRows; ++xx) {
+    for (int yy = 0; yy < kNumCols; ++yy) {
       // Get the number of degrees of freedom for the robot
       auto num_dofs = plant.num_actuated_dofs(models[xx][yy]);
       // Create a vector with the same number of zeros
